+++ pysteps/utils/__init__.py
@@ -1,9 +1,12 @@
 """Miscellaneous utility functions."""
 
 from .arrays import *
+from .cleansing import *
 from .conversion import *
 from .dimension import *
+from .images import *
 from .interface import get_method
+from .interpolate import *
 from .fft import *
 from .spectral import *
 from .transformation import *
